update readme with new rakudo version. annoyingly the latest (2020.01) won't compile HTTP:UserAgent
Actually this commit is a bit of a grab bag - the new version supports atomicint so using that
instead of the bodged up Channel-based lock

it also has a bit of death insurance in the main loop. and a slightly fiddled Pas::Store which
resulted from a first look at doing more with scripts - particularly storing them internally.

also a bit of help improvement