import command
required some jiggery pokery, leaving a few things looking like they
want a big tidy up - notably ASClient

also, unfortunately requred a fiddle with:
http-useragent/lib/HTTP/Request.pm6

at line 188 it was insisting on encoding the file content in ascii
in order to calculate the content length. unforch this blows up on
utf8 content, so i fiddled it to use the default encoding which
happens to be utf8. sigh.